+++ ros_ws/src/simulation/racer_sensors/src/main_racer_odometry.cpp
@@ -1,0 +1,9 @@
+#include "racer_odometry.h"
+
+int main(int argc, char** argv)
+{
+    ros::init(argc, argv, "racer_odometry");
+    RacerOdometry racerOdometry;
+    ros::spin();
+    return EXIT_SUCCESS;
+}
