+++ ros_ws/src/simulation/racer_sensors/src/racer_odometry.cpp
@@ -16,7 +16,7 @@ RacerOdometry::RacerOdometry()
 
 void RacerOdometry::robotPoseUpdate(const gazebo_msgs::LinkStates::ConstPtr& links)
 {
-    for (uint i = 0; i < links->name.size(); i++)
+    for (unsigned int i = 0; i < links->name.size(); i++)
     {
         if (links->name.at(i) == BASE_LINK_NAME)
         {
@@ -31,18 +31,10 @@ void RacerOdometry::robotPoseUpdate(const gazebo_msgs::LinkStates::ConstPtr& lin
 void RacerOdometry::timerCallback(const ros::TimerEvent&)
 {
     nav_msgs::Odometry msg;
-    msg.header.stamp = m_last_received_stamp;
-    msg.header.frame_id = "map";
-    msg.child_frame_id = "odom";
-    msg.pose.pose = m_last_received_pose;
-    msg.twist.twist = m_last_received_twist;
+    {
+        msg.header.stamp = m_last_received_stamp;
+        msg.pose.pose = m_last_received_pose;
+        msg.twist.twist = m_last_received_twist;
+    }
     m_odom_publisher.publish(msg);
-}
-
-int main(int argc, char** argv)
-{
-    ros::init(argc, argv, "racer_odometry");
-    RacerOdometry racerOdometry;
-    ros::spin();
-    return EXIT_SUCCESS;
 }
